pyproject.toml
[tool.poetry]
name = "ecommerce"
version = "0.1.0"
description = ""
authors = [""]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
selene = "2.0.0rc9"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
faker = "^33.1.0"
allure-pytest = "^2.13.5"
pytest = "^8.3.4"
pydantic-settings = "^2.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
